+++ http/Response.h
@@ -31,7 +31,7 @@ public:
     
     void set(const std::string& field, const std::string& value) const;
     
-    void cookie(const std::string& name, const std::string& value, const ResponseCookieOptions& options) const;
+    void cookie ( const std::string& name, const std::string& value, const ResponseCookieOptions& options ) const;
     
     void redirect(const std::string& name) const;
